Push forward with integrating & leverating the Polygon lib tools.
Hole cutouts are now created as polygons and subtracted from the original
part.

Much work on airfoil trailing edge stock cutouts (for standard stock sizes.)
This code does a "best fit" match that preserves the exact tip of the
original trailing edge and minimizes error for the remainder of the airfoil.
Also added an option to shave (or add) a wedge from the midpoint back to
make a perfect fit with stock trailing edge sizes.  This is an engineering
cheat, but allows a clean build with available materials.  It is optional
if the designer chooses to use it.